refactor: moved context to properly folder
