+++ src/presentation/components/Player/controls/index.tsx
@@ -2,8 +2,7 @@ import React from "react";
 import { Feather, FontAwesome5 } from "@expo/vector-icons";
 import { Pressable, View } from "react-native";
 import { Theme } from "@/presentation/styles";
-import { PlayerState } from "@/presentation/components";
-import { usePlayerContext } from "@/presentation/components";
+import { usePlayerContext, PlayerState } from "@/presentation/contexts";
 import styles from "../styles";
 
 interface ControlsProps {}
